+++ apps/shared/function.cpp
@@ -68,7 +68,7 @@ int Function::nameWithArgument(char * buffer, size_t bufferSize, CodePoint arg) 
   assert(UTF8Decoder::CharSizeOfCodePoint(arg) == 1);
   const char * functionName = fullName();
   size_t baseNameLength = SymbolAbstract::TruncateExtension(buffer, functionName, bufferSize - k_parenthesedArgumentLength);
-  int result = baseNameLength + strlcpy(&buffer[baseNameLength], k_parenthesedArgument, bufferSize-baseNameLength);
+  size_t result = baseNameLength + strlcpy(&buffer[baseNameLength], k_parenthesedArgument, bufferSize-baseNameLength);
   int bufferRemainingSize = bufferSize - (baseNameLength+1);
   if (bufferRemainingSize > 0) {
     UTF8Decoder::CodePointToChars(arg, buffer+baseNameLength+1, bufferRemainingSize);
